viejos/requirements.txt
pdfminer.six
elasticsearch
chardet
ocrmypdf
tesseract
